# Optimize blobstore ReadBlobs