--- a/server/backends/blobstore/aws/aws.go
+++ b/server/backends/blobstore/aws/aws.go
@@ -178,8 +178,17 @@ func (a *AwsS3BlobStore) createBucketIfNotExists(ctx context.Context, bucketName
 func (a *AwsS3BlobStore) ReadBlob(ctx context.Context, blobName string) ([]byte, error) {
 	start := time.Now()
 	b, err := a.download(ctx, blobName)
-	util.RecordReadMetrics(awsS3Label, start, b, err)
-	return util.Decompress(b, err)
+	duration := time.Since(start)
+	if err != nil {
+		return b, err
+	}
+	var buf bytes.Buffer
+	if _, err := buf.Write(b); err != nil {
+		return b, err
+	}
+	bytes, err := util.Decompress(&buf)
+	util.RecordReadMetrics(awsS3Label, duration, int64(len(b)), err)
+	return bytes, err
 }
 
 func (a *AwsS3BlobStore) download(ctx context.Context, blobName string) ([]byte, error) {