--- a/server/backends/blobstore/aws/aws.go
+++ b/server/backends/blobstore/aws/aws.go
@@ -179,15 +179,15 @@ func (a *AwsS3BlobStore) ReadBlob(ctx context.Context, blobName string) ([]byte,
 	start := time.Now()
 	b, err := a.download(ctx, blobName)
 	duration := time.Since(start)
+	// We update the err later on, so we need to wrap RecordReadMetrics in a func.
+	defer func() {
+		util.RecordReadMetrics(awsS3Label, duration, int64(len(b)), err)
+	}()
 	if err != nil {
 		return b, err
 	}
-	var buf bytes.Buffer
-	if _, err := buf.Write(b); err != nil {
-		return b, err
-	}
-	bytes, err := util.Decompress(&buf)
-	util.RecordReadMetrics(awsS3Label, duration, int64(len(b)), err)
+	rd := bytes.NewReader(b)
+	bytes, err := util.Decompress(rd)
 	return bytes, err
 }