--- a/server/backends/blobstore/aws/aws.go
+++ b/server/backends/blobstore/aws/aws.go
@@ -175,20 +175,18 @@ func (a *AwsS3BlobStore) createBucketIfNotExists(ctx context.Context, bucketName
 	return nil
 }
 
-func (a *AwsS3BlobStore) ReadBlob(ctx context.Context, blobName string) ([]byte, error) {
+func (a *AwsS3BlobStore) ReadBlob(ctx context.Context, blobName string) (rb []byte, retErr error) {
 	start := time.Now()
 	b, err := a.download(ctx, blobName)
 	duration := time.Since(start)
-	// We update the err later on, so we need to wrap RecordReadMetrics in a func.
 	defer func() {
-		util.RecordReadMetrics(awsS3Label, duration, int64(len(b)), err)
+		util.RecordReadMetrics(awsS3Label, duration, int64(len(b)), retErr)
 	}()
 	if err != nil {
 		return b, err
 	}
 	rd := bytes.NewReader(b)
-	bytes, err := util.Decompress(rd)
-	return bytes, err
+	return util.Decompress(rd)
 }
 
 func (a *AwsS3BlobStore) download(ctx context.Context, blobName string) ([]byte, error) {