--- a/server/backends/blobstore/disk/BUILD
+++ b/server/backends/blobstore/disk/BUILD
@@ -11,6 +11,7 @@ go_library(
         "//server/util/disk",
         "//server/util/ioutil",
         "//server/util/log",
+        "//server/util/status",
         "//server/util/tracing",
     ],
 )