--- a/server/backends/blobstore/util/BUILD
+++ b/server/backends/blobstore/util/BUILD
@@ -8,6 +8,7 @@ go_library(
     deps = [
         "//server/interfaces",
         "//server/metrics",
+        "//server/util/ioutil",
         "@com_github_prometheus_client_golang//prometheus",
         "@org_golang_google_grpc//status",
     ],